Extract core domain package
