+++ packages/core_domain/README.md
@@ -1,0 +1,7 @@
+# core_domain
+
+Shared pure-Dart domain primitives for the Flutter starter template.
+
+This package contains framework-agnostic contracts used across features:
+`Failure`, `Result`, and `UseCase`. Keep Flutter widgets, platform services,
+DI, analytics, and app-specific domain entities outside this package.
