+++ packages/core_domain/pubspec.yaml
@@ -1,0 +1,9 @@
+name: core_domain
+description: Shared pure-Dart domain primitives for the Flutter starter template.
+version: 0.1.0
+publish_to: none
+
+environment:
+  sdk: ^3.12.0
+
+resolution: workspace
